refactor: split content resources into seperate modules
