+++ src/posit/connect/content/__init__.py
@@ -1,3 +1,5 @@
-from .content import Content, ContentOwner, Contents
+from .content import Content
+from .content_owner import ContentOwner
+from .contents import Contents
 
 __all__ = ["Content", "ContentOwner", "Contents"]
